[refactor] 피드백 반영중 (2)
