+++ src/main/java/lotto/service/LottoGenerator.java
@@ -1,0 +1,9 @@
+package lotto.service;
+
+import lotto.Money;
+import lotto.model.Lottos;
+
+public interface LottoGenerator {
+
+    public Lottos generate(final Money money);
+}
